+++ apps/app/src/app/api/status/route.ts
@@ -8,11 +8,3 @@ import database from '@/lib/database';
 export async function GET() {
 	return NextResponse.json({ message: 'hi' });
 }
-
-// export async function POST() {
-// 	const db = await database('status', 'status');
-
-// 	console.log(db);
-
-// 	return NextResponse.json({ message: 'hi' });
-// }