+++ packages/lib/index.ts
@@ -4,3 +4,4 @@ export { auth } from './src/auth';
 export { navigation } from './src/navigation';
 export { proxy } from './src/proxy';
 export * from './src/env';
+export * from './src/products';