+++ packages/lib/src/products.ts
@@ -1,0 +1,5 @@
+export const products = {
+	servers: [],
+	users: ['prod_V0pSuzpbmBFsRv', 'prod_V0pRf7S9w9Pjb9', 'prod_V0pREm72CzuoVc', 'prod_V0pQ0EDH0iEDIm', 'prod_V0pD5uXoTQkrYS'],
+	developers: ['prod_V0oBbO5pVQMxcC'],
+};