+++ apps/api/CHANGELOG.md
@@ -1,5 +1,18 @@
 # api
 
+## 0.3.0
+
+### Minor Changes
+
+- Stripe addition
+
+### Patch Changes
+
+- Updated dependencies
+    - @xernerx/lib@0.3.0
+    - @xernerx/components@0.2.6
+    - @xernerx/providers@0.2.6
+
 ## 0.2.3
 
 ### Patch Changes