+++ packages/components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xernerx/components
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+    - @xernerx/lib@0.3.0
+    - @xernerx/providers@0.2.6
+
 ## 0.2.5
 
 ### Patch Changes